requirements.txt
lxml==4.4.0
Pillow==6.1.0
python-pptx==0.6.18
XlsxWriter==1.1.8
